chragetest/vnc-mcp-server/requirements.txt
mcp>=1.0.0
vncdotool>=1.2.0
Pillow>=10.0.0
